Minutes Before the New Year
New Year is coming and you are excited to know how many minutes remain before the New Year. You know that currently the clock shows h hours and m minutes, where 0≤hh<24 and 0≤mm<60. We use 24-hour time format!

Your task is to find the number of minutes before the New Year. You know that New Year comes when the clock shows 0 hours and 0 minutes.

You have to answer t independent test cases.

Input
The first line of the input contains one integer t (1≤t≤1439) — the number of test cases.

The following t lines describe test cases. The i-th line contains the time as two integers h and m (0≤h<24, 0≤m<60). It is guaranteed that this time is not a midnight, i.e. the following two conditions can't be met at the same time: h=0 and m=0. It is guaranteed that both h and m are given without leading zeros.

Output
For each test case, print the answer on it — the number of minutes before the New Year.


Example :
Input :
5
23 55
23 0
0 1
4 20
23 59

Output :
5
60
1439
1180
1